MPI seems to work

+++ cemc/wanglandau/ce_calculator.py
@@ -22,7 +22,7 @@ except Exception as exc:
     print (str(exc))
     print ("Could not find C++ version, falling back to Python version")
 
-def get_ce_calc( small_bc, bc_kwargs, eci, size=[1,1,1] ):
+def get_ce_calc( small_bc, bc_kwargs, eci=None, size=[1,1,1] ):
     """
     Constructs a CE calculator by first computing the correlation function
     from a small cell
@@ -34,27 +34,36 @@ def get_ce_calc( small_bc, bc_kwargs, eci, size=[1,1,1] ):
     eci - Effective Cluster Interactions
     size - The atoms in small_bc will be extended by this amount
     """
-    calc1 = CE( small_bc, eci )
-    init_cf = calc1.get_cf()
-    cell_lenghts = small_bc.atoms.get_cell_lengths_and_angles()[:3]
-    min_length = np.min(cell_lenghts)/2.0
-
-    bc_kwargs["size"] = size
-    bc_kwargs["max_cluster_dia"] = min_length
-
     rank = MPI.COMM_WORLD.Get_rank()
-    db_name = "temporary_db_{}.db".format(rank)
-    if ( os.path.exists(db_name) ):
-        os.remove( db_name )
-    bc_kwargs["db_name"] = db_name
+    unknown_type = False
+    large_bc = small_bc # Just for the other processes
+    init_cf = None
+    if ( rank == 0 ):
+        calc1 = CE( small_bc, eci )
+        init_cf = calc1.get_cf()
+        cell_lenghts = small_bc.atoms.get_cell_lengths_and_angles()[:3]
+        min_length = np.min(cell_lenghts)/2.0
 
-    if ( isinstance(small_bc,BulkCrystal) ):
-        large_bc = BulkCrystal(**bc_kwargs)
-        print ("here")
-    elif ( isinstance(small_bc,BulkSpacegroup) ):
-        large_bc = BulkSpacegroup(**bc_kwargs)
-    else:
+        bc_kwargs["size"] = size
+        bc_kwargs["max_cluster_dia"] = min_length
+
+        db_name = "temporary_db.db"
+        if ( os.path.exists(db_name) ):
+            os.remove( db_name )
+        bc_kwargs["db_name"] = db_name
+
+        if ( isinstance(small_bc,BulkCrystal) ):
+            large_bc = BulkCrystal(**bc_kwargs)
+        elif ( isinstance(small_bc,BulkSpacegroup) ):
+            large_bc = BulkSpacegroup(**bc_kwargs)
+        else:
+            unknown_type = True
+
+    unknown_type = MPI.COMM_WORLD.bcast( unknown_type, root=0 )
+    if ( unknown_type ):
         raise TypeError( "The small_bc argument has to by of type BulkCrystal or BulkSpacegroup" )
+    large_bc = MPI.COMM_WORLD.bcast( large_bc, root=0 )
+    init_cf = MPI.COMM_WORLD.bcast( init_cf, root=0 )
     calc2 = CE( large_bc, eci, initial_cf=init_cf )
     return calc2
 
@@ -64,15 +73,18 @@ class CE( Calculator ):
     """
 
     implemented_properties = ["energy"]
-    def __init__( self, BC, eci, initial_cf=None ):
+    def __init__( self, BC, eci=None, initial_cf=None ):
         Calculator.__init__( self )
         self.BC = BC
-        self.eci = eci
         self.corrFunc = CorrFunction(self.BC)
         if ( initial_cf is None ):
             self.cf = self.corrFunc.get_cf( self.BC.atoms )
         else:
             self.cf = initial_cf
+
+        if ( eci is None ):
+            eci = {name:1.0 for name in self.cf.keys()}
+        self.eci = eci
         # Make supercell
         self.atoms = self.BC.atoms
         symbols = [atom.symbol for atom in self.BC.atoms] # Keep a copy of the original symbols
